+++ W3ChampionsStatisticService/Admin/Permissions/PermissionsController.cs
@@ -1,18 +1,17 @@
 ﻿using System.Threading.Tasks;
 using Microsoft.AspNetCore.Mvc;
-using W3ChampionsStatisticService.Ports;
 using W3ChampionsStatisticService.WebApi.ActionFilters;
 using System.Net.Http;
 using W3C.Contracts.Admin.Permission;
+using W3ChampionsStatisticService.Services;
 
 namespace W3ChampionsStatisticService.Admin.Permissions;
 
 [ApiController]
 [Route("api/admin/permissions")]
-public class PermissionsController(
-    IPermissionsRepository permissionsRepository) : ControllerBase
+public class PermissionsController(IdentityServiceClient identityServiceClient) : ControllerBase
 {
-    private readonly IPermissionsRepository _permissionsRepository = permissionsRepository;
+    private readonly IdentityServiceClient _identityServiceClient = identityServiceClient;
 
     [HttpGet]
     [InjectAuthToken]
@@ -20,7 +19,7 @@ public class PermissionsController(
     {
         try
         {
-            var permissions = await _permissionsRepository.GetPermissions(authToken);
+            var permissions = await _identityServiceClient.GetPermissions(authToken);
             return Ok(permissions);
         }
         catch (HttpRequestException ex)
@@ -35,7 +34,7 @@ public class PermissionsController(
     {
         try
         {
-            await _permissionsRepository.AddAdmin(permission, authToken);
+            await _identityServiceClient.AddAdmin(permission, authToken);
             return Ok();
         }
         catch (HttpRequestException ex)
@@ -50,7 +49,7 @@ public class PermissionsController(
     {
         try
         {
-            await _permissionsRepository.EditAdmin(permission, authToken);
+            await _identityServiceClient.EditAdmin(permission, authToken);
             return Ok();
         }
         catch (HttpRequestException ex)
@@ -65,7 +64,7 @@ public class PermissionsController(
     {
         try
         {
-            await _permissionsRepository.DeleteAdmin(id, authToken);
+            await _identityServiceClient.DeleteAdmin(id, authToken);
             return Ok();
         }
         catch (HttpRequestException ex)
